+++ .github/workflows/aws_tests.yml
@@ -39,5 +39,9 @@ jobs:
       - name: Install full requirements
         run: pip install .[aws,test]
 
+      # TODO: when passing, reverse order of tests from minimal to data
+      - name: Run subset of tests that use S3 live services
+        run: pytest -rsx -n auto tests/test_on_data/test_yaml/yaml_aws_tools.py
+
       - name: Run subset of tests that use S3 live services
         run: pytest -rsx -n auto tests/test_minimal/test_tools/aws_tools.py
